* 6.2:
  [Clock] fix usleep deprecation warning
  [Cache] sync Redis6Proxy with upstream
  [HttpFoundation] IPv4-mapped IPv6 addresses incorrectly rejected
  Minor (comment only), part 2
  [RateLimiter] Add typecast to Reservation::wait
  chore: fix syntax error
  [FrameworkBundle] add `kernel.locale_aware` tag to `LocaleSwitcher`
  [FrameworkBundle] Remove check of undefined service in mailer assertions
  [VarDumper] Ignore \Error in __debugInfo()

+++ src/Symfony/Component/Clock/MonotonicClock.php
@@ -67,7 +67,7 @@ final class MonotonicClock implements ClockInterface
         }
 
         if (0 < $us = $seconds - $s) {
-            usleep($us * 1E6);
+            usleep((int) ($us * 1E6));
         }
     }
